# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Manabu Japanese Learning Application",
        "title": "Manabu API - Japanese Learning Application",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.manabu.com/support",
            "email": "support@manabu.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/api/v1",
    "paths": {
...
        "/vocabularies": {
            "get": {
                "description": "Retrieve vocabularies with advanced filtering, search, sorting, and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabularies"
                ],
                "summary": "Get all Vocabularies",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 5,
                        "description": "Filter by JLPT Level ID",
                        "name": "jlptLevelId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Filter by Category ID",
                        "name": "categoryId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"noun\"",
                        "description": "Filter by part of speech",
                        "name": "partOfSpeech",
                        "in": "query"
                    },
                    {
                        "maximum": 5,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Filter by difficulty (1-5)",
                        "name": "difficulty",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"dog\"",
                        "description": "Search in word, reading, or meaning",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at",
                        "example": "\"word\"",
                        "description": "Sort by field (word, difficulty, created_at)",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "example": "\"asc\"",
                        "description": "Sort order (asc, desc)",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.VocabularyListSwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new vocabulary entry (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabularies"
                ],
                "summary": "Create Vocabulary",
                "parameters": [
                    {
                        "description": "Vocabulary details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateVocabularyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.VocabularySwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "Vocabulary already exists for this JLPT level",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "422": {
                        "description": "Invalid JLPT level ID or Category ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/vocabularies/{id}": {
            "get": {
                "description": "Retrieve a specific vocabulary entry by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabularies"
                ],
                "summary": "Get Vocabulary by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vocabulary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.VocabularySwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Vocabulary not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing vocabulary entry by ID (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabularies"
                ],
                "summary": "Update Vocabulary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vocabulary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated vocabulary details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateVocabularyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.VocabularySwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Vocabulary not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "Vocabulary already exists for this JLPT level",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "422": {
                        "description": "Invalid JLPT level ID or Category ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a vocabulary entry by ID (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabularies"
                ],
                "summary": "Delete Vocabulary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vocabulary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Vocabulary not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Category for animal-related vocabulary"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "jlptLevel": {
                    "$ref": "#/definitions/dto.JlptLevelResponse"
                },
                "jlptLevelId": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Animals"
                }
            }
        },
...
        "dto.CreateVocabularyRequest": {
            "type": "object",
            "required": [
                "categoryId",
                "jlptLevelId",
                "meaning",
                "word"
            ],
            "properties": {
                "audioUrl": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "https://example.com/audio/inu.mp3"
                },
                "categoryId": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "difficulty": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1,
                    "example": 1
                },
                "exampleSentence": {
                    "type": "string",
                    "example": "犬が好きです"
                },
                "exampleSentenceMeaning": {
                    "type": "string",
                    "example": "I like dogs"
                },
                "exampleSentenceReading": {
                    "type": "string",
                    "example": "いぬがすきです"
                },
                "imageUrl": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "https://example.com/images/dog.jpg"
                },
                "jlptLevelId": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 5
                },
                "meaning": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1,
                    "example": "dog"
                },
                "partOfSpeech": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "noun"
                },
                "reading": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "いぬ"
                },
                "word": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "犬"
                }
            }
        },
        "dto.JlptLevelResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "N5"
                },
                "description": {
                    "type": "string",
                    "example": "Basic level"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "levelOrder": {
                    "type": "integer",
                    "example": 5
                },
                "name": {
                    "type": "string",
                    "example": "JLPT N5"
                }
            }
        },
...
        "dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "totalItems": {
                    "type": "integer",
                    "example": 50
                },
                "totalPages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "dto.UpdateVocabularyRequest": {
            "type": "object",
            "required": [
                "categoryId",
                "jlptLevelId",
                "meaning",
                "word"
            ],
            "properties": {
                "audioUrl": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "https://example.com/audio/inu.mp3"
                },
                "categoryId": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "difficulty": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1,
                    "example": 1
                },
                "exampleSentence": {
                    "type": "string",
                    "example": "犬が好きです"
                },
                "exampleSentenceMeaning": {
                    "type": "string",
                    "example": "I like dogs"
                },
                "exampleSentenceReading": {
                    "type": "string",
                    "example": "いぬがすきです"
                },
                "imageUrl": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "https://example.com/images/dog.jpg"
                },
                "jlptLevelId": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 5
                },
                "meaning": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1,
                    "example": "dog"
                },
                "partOfSpeech": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "noun"
                },
                "reading": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "いぬ"
                },
                "word": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "犬"
                }
            }
        },
...
        "dto.VocabularyListSwaggerResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.VocabularyResponse"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Vocabularies retrieved successfully"
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationResponse"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "dto.VocabularyResponse": {
            "type": "object",
            "properties": {
                "audioUrl": {
                    "type": "string",
                    "example": "https://example.com/audio/inu.mp3"
                },
                "category": {
                    "$ref": "#/definitions/dto.CategoryResponse"
                },
                "categoryId": {
                    "type": "integer",
                    "example": 1
                },
                "difficulty": {
                    "type": "integer",
                    "example": 1
                },
                "exampleSentence": {
                    "type": "string",
                    "example": "犬が好きです"
                },
                "exampleSentenceMeaning": {
                    "type": "string",
                    "example": "I like dogs"
                },
                "exampleSentenceReading": {
                    "type": "string",
                    "example": "いぬがすきです"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "imageUrl": {
                    "type": "string",
                    "example": "https://example.com/images/dog.jpg"
                },
                "jlptLevel": {
                    "$ref": "#/definitions/dto.JlptLevelResponse"
                },
                "jlptLevelId": {
                    "type": "integer",
                    "example": 5
                },
                "meaning": {
                    "type": "string",
                    "example": "dog"
                },
                "partOfSpeech": {
                    "type": "string",
                    "example": "noun"
                },
                "reading": {
                    "type": "string",
                    "example": "いぬ"
                },
                "word": {
                    "type": "string",
                    "example": "犬"
                }
            }
        },
        "dto.VocabularySwaggerResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.VocabularyResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Vocabulary created successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {},
                "status": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}